update the database, and api - add address wishlidt search

+++ BackEnd/api/v1/views/cartItem.py
@@ -8,8 +8,8 @@ from models.cart import CartItem
 @app_views.route("/cartItems", methods=['GET'], strict_slashes=False)
 def cartItems():
     """get store data"""
-    if request.user and request.user.cart:
-        cartItems = request.user.cart.cartItems
+    if request.user:
+        cartItems = request.user.cartItems
         return jsonify({"cartItems": [crtItm.to_dict() for crtItm in cartItems]}), 200
     abort(404)
 
@@ -19,20 +19,19 @@ def addCart(product_id):
     """get store data"""
     data = request.get_json(force=True, silent=True)
     if not data:
-        quantity = 1
+        data['quantity'] = 1
 
     product = storage.get(Product, product_id)
-    if product:
-        if request.user and request.user.cart:
-            if f"'product_id': {product_id}" in f"{[crtItm.to_dict() for crtItm in request.user.cart.cartItems]}":
-                return jsonify({'error': 'you have that product in carts'})
-            crtItem = CartItem(
-                cart_id=request.user.cart.id,
-                product_id=product_id,
-                quantity=data['quantity'] if data.get('quantity') else quantity
-            )
-            crtItem.save()
-            return jsonify({"message": "product added successfully"}), 200
+    if product and request.user:
+        if f"'product_id': {product_id}" in f"{[crtItm.to_dict() for crtItm in request.user.cartItems]}":
+            return jsonify({'error': 'you have that product in carts'})
+        crtItem = CartItem(
+            user_id=request.user.id,
+            product_id=product_id,
+            quantity=data['quantity']
+        )
+        crtItem.save()
+        return jsonify({"message": "cart added successfully"}), 200
     abort(404)
 
 
@@ -43,9 +42,9 @@ def updateCart(cartItem_id):
     if not data:
         return jsonify({"error": "check your data send!"}), 400
 
-    if request.user and request.user.cart:
+    if request.user:
         cartItem = storage.get(CartItem, cartItem_id)
-        if cartItem:
+        if cartItem and cartItem in request.user.cartItems:
             storage.update(cartItem, **data)
             storage.save()
             return jsonify({"message": "cart updated successfully"}), 200
@@ -55,9 +54,9 @@ def updateCart(cartItem_id):
 @app_views.route("/cartItems/<int:cartItem_id>", methods=['DELETE'], strict_slashes=False)
 def deleteCart(cartItem_id):
     """delete store"""
-    if request.user and request.user.cart:
+    if request.user:
         cartItem = storage.get(CartItem, cartItem_id)
-        if cartItem:
+        if cartItem and cartItem in request.user.cartItems:
             cartItem.delete()
             storage.save()
             return jsonify({"message": "cart deleted successfully"}), 200
